File for NodeStatus
